fully implemented Registration screen and Login Screen, created a common styles sheet, general code review for Registration screen and Login Screen
