+++ App.tsx
@@ -13,7 +13,7 @@ import MapScreen from './src/screens/MapScreen';
 import TicketScreen from './src/screens/TicketsScreen';
 import ProfileScreen from './src/screens/ProfileScreen';
 import LoginScreen from './src/screens/Login/LoginScreen';
-import RegisterScreen from './src/screens/Login/RegisterScreen';
+import RegisterScreen from '@/screens/Login/registerScreen';
 import ForgotPasswordScreen from '@/screens/Login/ForgotPassword';
 import OwnerLoginScreen from '@/screens/Login/OwnerLogScreen';
